unit tests and refactoring
