+++ inner-tool/src/main/java/common/GenCodeRequest.java
@@ -1,10 +1,5 @@
-package util;
+package common;
 
-import data.Settings;
-
-/**
- * @see message.GenCodeMessage
- */
 public class GenCodeRequest {
     private Settings settings = new Settings();
