+++ src/main/java/ui/dialog/TextEditorComponent.java
@@ -10,8 +10,7 @@ import com.intellij.openapi.editor.ex.EditorEx;
 import com.intellij.openapi.fileTypes.FileType;
 import com.intellij.openapi.fileTypes.FileTypeManager;
 import com.intellij.openapi.project.Project;
-import com.intellij.openapi.project.ProjectManager;
-import data.Settings;
+import common.Settings;
 
 import javax.swing.*;
 import java.awt.*;
