+++ src/main/java/util/debugger/PluginSaveLoader.java
@@ -11,11 +11,11 @@ import com.sun.jdi.ObjectReference;
 import com.sun.jdi.StringReference;
 import com.sun.jdi.ThreadReference;
 import com.sun.jdi.Value;
-import data.Settings;
+import common.Settings;
 import data.VariableInfo;
 import data.model.SavedValue;
 import data.tool.GenCodeMessage;
-import data.tool.GenCodeRequest;
+import common.GenCodeRequest;
 import data.tool.ToolMessage;
 import util.exception.JsonSerializeException;
 import util.exception.SaveValueInnerException;
